Move more stats classes around
