+++ feature/stats/src/main/java/com/github/saikcaskey/pokertracker/stats/domain/components/factory/StatsComponentFactory.kt
@@ -1,4 +1,4 @@
-package com.github.saikcaskey.pokertracker.stats.factory
+package com.github.saikcaskey.pokertracker.stats.domain.components.factory
 
 import com.github.saikcaskey.pokertracker.domain.presentation.component.factory.ChildComponentFactory
 import com.github.saikcaskey.pokertracker.stats.presentation.navigation.StatsPagerNavigationRoute
