+++ feature/stats/src/main/java/com/github/saikcaskey/pokertracker/stats/presentation/components/StatsFeaturePagerComponentImpl.kt
@@ -10,7 +10,7 @@ import com.arkivanov.decompose.value.Value
 import com.arkivanov.decompose.value.subscribe
 import com.github.saikcaskey.pokertracker.domain.presentation.component.FeatureComponent
 import com.github.saikcaskey.pokertracker.stats.domain.components.StatsFeaturePagerComponent
-import com.github.saikcaskey.pokertracker.stats.factory.StatsComponentFactory
+import com.github.saikcaskey.pokertracker.stats.domain.components.factory.StatsComponentFactory
 import com.github.saikcaskey.pokertracker.stats.presentation.navigation.StatsPagerNavigationRoute
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
